+++ Source/Singulink.Collections/HashSetDictionary.ValueSet.cs
@@ -9,7 +9,7 @@ namespace Singulink.Collections;
 public partial class HashSetDictionary<TKey, TValue>
 {
     /// <summary>
-    /// Represents a dictionary synchronized set of values associcated with a key in a <see cref="HashSetDictionary{TKey, TValue}"/>.
+    /// Represents a dictionary synchronized set of values associated with a key in a <see cref="HashSetDictionary{TKey, TValue}"/>.
     /// </summary>
     public sealed class ValueSet : ReadOnlyValueSet, ISet<TValue>, IReadOnlyCollectionProvider<TValue>
     {